Tests: Always use host retry_interval
Currently, when a service is in a HARD CRITICAL it instantly trigges a
host-check. As a result the host retry_interval is basicly useless when
a host (and presumebly therefore its services) is down.

This patch adds a test to ensure that a host uses its own retry_interval
even when it has a service in HARD CRITICAL.

This is part of MON-5625

Signed-off-by: Jacob Hansen <jhansen@op5.com>