Use cached host state on service state changes
Instead of triggering a new host check every time a state changes, we
instead use the chached host state. This ensures that a host is checked
as per its configured retry_interval even if it has a service which is
in HARD CRITICAL.

This is part of MON-5625

This is an Nagios patch by Frederic Van Espen <frederic.ve@gmail.com>
which has been ported to Naemon.

Co-authored-by: Frederic Van Espen <frederic.ve@gmail.com>
Signed-off-by: Jacob Hansen <jhansen@op5.com>